+++ src/Components/QuizList.jsx
@@ -2,6 +2,7 @@ import React, { useEffect, useState } from 'react';
 import axios from 'axios';
 import { useNavigate } from 'react-router-dom';
 import '../Styles/QuizList.css';
+import Swal from 'sweetalert2';
 
 const QuizList = () => {
   const [quizzes, setQuizzes] = useState([]);
@@ -24,16 +25,29 @@ const QuizList = () => {
     fetchQuizzesAndSubjects();
   }, []);
 
-  const handleDelete = async (id) => {
-    try {
-      await axios.delete(`http://localhost:8080/api/quizzes/delete-quiz/${id}`);
-      alert('Quiz deleted successfully!');
-      setQuizzes(quizzes.filter((quiz) => quiz.id !== id));
-    } catch (error) {
-      console.error('There was an error deleting the quiz!', error);
-      alert('There was an error deleting the quiz!');
+const handleDelete = async(id) => {
+  Swal.fire({
+    title: 'Are you sure?',
+    text: "This action will delete the quiz permanently!",
+    icon: 'warning',
+    showCancelButton: true,
+    confirmButtonText: 'Yes, delete it!',
+    cancelButtonText: 'Cancel'
+  }).then(async (result) => {
+    if (result.isConfirmed) {
+      try {
+        await axios.delete(`http://localhost:8080/api/quizzes/delete-quiz/${id}`);
+        Swal.fire('Deleted!', 'Quiz has been deleted.', 'success').then(() => {
+          setQuizzes(quizzes.filter((quiz) => quiz.id !== id));
+        });
+      } catch (error) {
+        Swal.fire('Error', 'There was an error deleting the quiz!', 'error');
+        console.error('Error deleting the quiz', error);
+      }
     }
-  };
+  });
+};
+
 
   const handleUpdate = (id) => {
     navigate(`/admin/update-quiz/${id}`);
@@ -53,7 +67,6 @@ const QuizList = () => {
           <li className="quiz-item" key={quiz.id}>
             <h3 className="quiz-question">{quiz.question}</h3>
             <p className="quiz-detail"><strong>Subject:</strong> {Number.isNaN(parseInt(quiz.subject, 10)) ? quiz.subject : subjectNameById(parseInt(quiz.subject))}</p>
-            <p className="quiz-detail"><strong>Type:</strong> {quiz.questionType}</p>
             <p className="quiz-detail"><strong>Choices:</strong> {quiz.choices.join(', ')}</p>
             <p className="quiz-detail"><strong>Correct Answer:</strong> {quiz.correctChoice.join(', ')}</p>
             <button className="btn delete" onClick={() => handleDelete(quiz.id)}>Delete</button>
